+++ Led.h
@@ -5,10 +5,10 @@ class Led {
     int pin;
 //    int onTime[100] = {};
 //    int offTime[100] = {};
-    int onTime = 0;
-    int offTime = 0;
+    unsigned long onTime = 0;
+    unsigned long offTime = 0;
     int lastUsedIndex = -1;
-    int delayTime = 1400;
+    unsigned long delayTime = (unsigned long) 1400;
   public:
     Led(int ledPin);
   	
@@ -19,9 +19,9 @@ class Led {
     
     int getPin();
     
-    void setOnTime(int onTime);
-    int getOnTime();
-    void setOffTime(int offTime);
-    int getOffTime();
-    int getDelayTime();
+    void setOnTime(unsigned long onTime);
+    unsigned long getOnTime();
+    void setOffTime(unsigned long offTime);
+    unsigned long getOffTime();
+    unsigned long getDelayTime();
 };
